+++ components/operations/common/stop-creation.tsx
@@ -1,0 +1,95 @@
+import * as z from 'zod';
+import {
+  BytesValue,
+  TypedValue,
+  ContractCallPayloadBuilder,
+  ContractFunction,
+} from '@multiversx/sdk-core';
+import { useForm } from 'react-hook-form';
+import { zodResolver } from '@hookform/resolvers/zod';
+import { Form } from '@/components/ui/form';
+import {
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+} from '@/components/ui/dialog';
+import {
+  commonOpertationsGasLimit,
+  builtInSC,
+} from '@/components/operations/constants';
+import { OperationsInputField } from '@/components/operations/operations-input-field';
+import { OperationsSubmitButton } from '@/components/operations/operations-submit-button';
+import { useContext } from 'react';
+import { OperationsStateDialogContext } from '@/components/operations/operations-status-dialog';
+import { OperationContentProps } from '@/components/operations/operations-common-types';
+
+const formSchema = z.object({
+  tokenId: z.string().min(1, 'The field is required!'),
+});
+
+export const StopCreation = ({ triggerTx, close }: OperationContentProps) => {
+  const { setOpen: setTxStatusDialogOpen } = useContext(
+    OperationsStateDialogContext
+  );
+
+  const form = useForm<z.infer<typeof formSchema>>({
+    resolver: zodResolver(formSchema),
+    defaultValues: {
+      tokenId: '',
+    },
+  });
+
+  const onSubmit = ({ tokenId }: z.infer<typeof formSchema>) => {
+    const args: TypedValue[] = [BytesValue.fromUTF8(tokenId.trim())];
+
+    const data = new ContractCallPayloadBuilder()
+      .setFunction(new ContractFunction('stopNFTCreate'))
+      .setArgs(args)
+      .build();
+
+    triggerTx?.({
+      address: builtInSC,
+      gasLimit: commonOpertationsGasLimit,
+      data,
+      value: 0,
+    });
+
+    setTxStatusDialogOpen(true);
+    form.reset();
+    close();
+  };
+
+  return (
+    <>
+      <DialogHeader className="p-8 pb-0">
+        <DialogTitle>Stop creation</DialogTitle>
+        <DialogDescription>
+          The ESDT manager can stop the creation of the token for the given ESDT
+          forever by removing the only ESDTRoleNFTCreate role available.
+        </DialogDescription>
+      </DialogHeader>
+      <div className="overflow-y-auto py-0 px-8">
+        <Form {...form}>
+          <form
+            id="stop-creation-form"
+            onSubmit={form.handleSubmit(onSubmit)}
+            className="space-y-8"
+          >
+            <div className="flex-1 overflow-auto p-1">
+              <OperationsInputField
+                name="tokenId"
+                label="Token id"
+                placeholder="Example: MyToken-23432"
+                description="Please provide your token id"
+              />
+            </div>
+          </form>
+        </Form>
+      </div>
+      <DialogFooter className="py-4 px-8">
+        <OperationsSubmitButton formId="stop-creation-form" />
+      </DialogFooter>
+    </>
+  );
+};