+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+### [0.2.0](https://github.com/xdevguild/buildo.dev/releases/tag/v0.2.0) (2023-10-22)
+- add transfer creation role for NFT/SFT/Meta
+- add stop creation for NFT/SFT/Meta
+- add transfer ownership for NFT/SFT/META
+- add check token data utility
+- update dependencies
+
 ### [0.1.0](https://github.com/xdevguild/buildo.dev/releases/tag/v0.1.0) (2023-10-14)
 - initial code